fix sorting bug
